Move DataProxy pojo file from manager-common to inlong-common.
