+++ inlong-common/src/main/java/org/apache/inlong/commons/pojo/dataproxy/DataProxyConfigResponse.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.manager.common.pojo.dataproxy;
+package org.apache.inlong.commons.pojo.dataproxy;
 
 /**
  * DataProxyConfigResponse
