+++ inlong-manager/manager-common/pom.xml
@@ -153,6 +153,11 @@
             </exclusions>
         </dependency>
 
+        <dependency>
+            <groupId>org.apache.inlong</groupId>
+            <artifactId>inlong-common</artifactId>
+            <version>${project.version}</version>
+        </dependency>
     </dependencies>
 
 </project>
