+++ inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/repository/DataProxyConfigRepository.java
@@ -28,17 +28,17 @@ import java.util.Timer;
 import java.util.TimerTask;
 
 import org.apache.commons.codec.digest.DigestUtils;
-import org.apache.inlong.manager.common.pojo.dataproxy.CacheClusterObject;
-import org.apache.inlong.manager.common.pojo.dataproxy.CacheClusterSetObject;
-import org.apache.inlong.manager.common.pojo.dataproxy.CacheTopicObject;
-import org.apache.inlong.manager.common.pojo.dataproxy.DataProxyCluster;
+import org.apache.inlong.commons.pojo.dataproxy.CacheClusterObject;
+import org.apache.inlong.commons.pojo.dataproxy.CacheClusterSetObject;
+import org.apache.inlong.commons.pojo.dataproxy.CacheTopicObject;
+import org.apache.inlong.commons.pojo.dataproxy.DataProxyCluster;
+import org.apache.inlong.commons.pojo.dataproxy.DataProxyConfigResponse;
+import org.apache.inlong.commons.pojo.dataproxy.InLongIdObject;
+import org.apache.inlong.commons.pojo.dataproxy.ProxyChannel;
+import org.apache.inlong.commons.pojo.dataproxy.ProxyClusterObject;
+import org.apache.inlong.commons.pojo.dataproxy.ProxySink;
+import org.apache.inlong.commons.pojo.dataproxy.ProxySource;
 import org.apache.inlong.manager.common.pojo.dataproxy.DataProxyClusterSet;
-import org.apache.inlong.manager.common.pojo.dataproxy.DataProxyConfigResponse;
-import org.apache.inlong.manager.common.pojo.dataproxy.InLongIdObject;
-import org.apache.inlong.manager.common.pojo.dataproxy.ProxyChannel;
-import org.apache.inlong.manager.common.pojo.dataproxy.ProxyClusterObject;
-import org.apache.inlong.manager.common.pojo.dataproxy.ProxySink;
-import org.apache.inlong.manager.common.pojo.dataproxy.ProxySource;
 import org.apache.inlong.manager.dao.entity.CacheCluster;
 import org.apache.inlong.manager.dao.entity.CacheClusterExt;
 import org.apache.inlong.manager.dao.entity.CacheTopic;
